graphcomponents rename + small cleanup
